+++ tests/test_schematron.py
@@ -3126,7 +3126,7 @@ class ElementCitationTests(PhaseBasedTestCase):
     def test_allowed_publication_types(self):
         for pub_type in ['journal', 'book', 'webpage', 'thesis', 'confproc',
                          'patent', 'software', 'database', 'legal-doc', 'newspaper',
-                         'other']:
+                         'other', 'report']:
             sample = u"""<article>
                           <back>
                             <ref-list>